Start to rename all code with the pbuilder* or pb_* prefix.
Start to rename funcs and structs to something more meaningful that
stands for parallel builder.

Remove the pg_node_building_st struct; its only useful field, avail_pos,
was added to the pbuilder_node_st struct as pool_pos that indicates the
thread position in the pool that is building the given node.

Code clean-up

Signed-off-by: Pedro Aguilar <paguilar.paguilar.org>