graph: Skip already built packages from a previous execution
If the previous exection was somehow halted, the current one starts
where the previous one left. It checks if the build directory of the
package exists; if yes, it also checks that it was installed. If not, it
builds the package.
This check may not work if a package does not install anything.

Signed-off-by: Pedro Aguilar <paguilar@paguilar.org>