graph_exec: Show the list of failed pkgs at the end of the build
When the execution terminal has a few lines of output buffer and there are many packages
building and one of them failed or the debug is enabled, the name of the failed pkg can be lost.

Let's display also the list of failed pkgs at the end of the process so they're not missed even
with just a few lines of output to the terminal.

Signed-off-by: Pedro Aguilar <paguilar@paguilar.org>