main.c: Unset BUILD_DIR environment variable
Some packages use this environment variable in a different manner, so
keeping it set causes problems.

For example, in the host-ztd package, if BUILD_DIR lib/Makefile is
defined, it installs the libs/binaries in the path specified by BUILD_DIR
resulting in all objects being installed in Buildroot's top-level
build directory instead of using the package's build directory.

Signed-off-by: Nolan Davenport
Signed-off-by: Pedro Aguilar <paguilar@paguilar.org>