graph_exec.c: Changed the top-level parallelization algorithm to one that maximizes the amount of packages being built in parallel at any given time. This further increases time savings.
The idea behind the new algorithm is to start the build for a package as soon as two conditions are met (give or take a second):
1. All of the package's dependencies were successfully built.
2. There is at least 1 thread available to create for the package to use for it's build.

I used the beaglebone_qt5 configuration for testing. With the previous br-pbuilder algorithm, I had 12% build time savings, and with the new algorithm, I had 23% build time savings.

The above test was performed with this hardware:
Ryzen 9 3900X 12 core 24 thread
NVME SSD
32GB DDR4 RAM

Signed-off-by: Nolan Davenport <nolanrdavenport@gmail.com>