Execute br2-external only once when invoking pbuilder
The br2-external script is executed every time that 'make' is invoked
and modifies the .br-external* files inside the building dir.
In the case of br-pbuilder, it's executed for every package and since
there can be N packages building at the same time, this script can be
called simultaneously changing those files and therefore esulting in
a concurrency issue in which the building process of any given package
results in the "No rule to make target <package>" error.

The changes in this commit make sure that only the first call to
br2-external is executed as expected, but all the following ones exit
immediately since they don't longer seem to be necessary while pbuilder
is in execution.
At the end of the build or when there is an error the file used as
flag is removed.

This should fix issue #4

Signed-off-by: Pedro Aguilar <paguilar@paguilar.org>